requirements.txt
OpenSearchRec
pandas
newspaper3k
langdetect
sentence-transformers
Flask
rapidfuzz
gunicorn